Update UserSerializer to allow modifications of preferred_name through first_name
We want to allow users to change their preferred_name without providing the preferred_name
field in the serializer to avoid the need to modify products already using this serializer.
Thus, the user would modify their preferred_name through entering a value in the first_name
field. The serializer would then take this value and save it to the user instance's preferred_name
field if the preferred_name is different from the first_name.

We previously had a SerializerMethodField for the first_name, but SerializerMethodFields are
read_only, preventing us from accessing first_name through validated_data. Thus, first_name
is changed to a CharField with source set to the get_preferred_name() method of the User model
with required=False so that the field could be left empty in a PUT request.