Remove get_phone_number() from PhoneNumberSerializer
get_phone_number() in the PhoneNumberSerializer uses a SerializerMethodField() to return the phone_number as a string instead of a PhoneNumber() object. However, we want the users to be able to change the phone_number and SerializerMethodFields are read_only. Therefore, remove the get_phone_number() method from the serializer and add a __str__ method in the PhoneNumberModel to return the string representation of the number.